requirements.txt
notepad requirements.txt
```

Paste this:
```
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.23.0
matplotlib>=3.6.0
scikit-learn>=1.2.0
gradio>=3.50.0